moved files to windows subdirectory, other small changes
