--- edk2toollib/uefi/telem/plugins/__init__.py
@@ -1,13 +1,0 @@
-# @file __init__.py
-# Imports all classes within this directory
-#
-# Copyright (c) Microsoft Corporation
-#
-# SPDX-License-Identifier: BSD-2-Clause-Patent
-##
-
-from os.path import dirname, basename, isfile
-import glob
-
-modules = glob.glob(dirname(__file__)+"/*.py")
-__all__ = [ basename(f)[:-3] for f in modules if isfile(f) and not basename(f).startswith('__')]
