+++ edk2toollib/windows/telem/decoders/__init__.py
@@ -1,0 +1,7 @@
+##
+# File to mark this a python package
+#
+# Copyright (c) Microsoft Corporation
+#
+# SPDX-License-Identifier: BSD-2-Clause-Patent
+##
