+++ edk2toollib/windows/telem/decoders/example_plugin.py
@@ -8,7 +8,7 @@
 
 import sys
 import uuid
-from edk2toollib.uefi.telem.cper_section_data import SECTION_PARSER_PLUGIN
+from edk2toollib.windows.telem.cper_section_data import SECTION_PARSER_PLUGIN
 
 class EXAMPLE_PLUGIN(SECTION_PARSER_PLUGIN):
