+++ edk2toollib/windows/telem/decoders/mu_plugin.py
@@ -9,7 +9,7 @@
 import sys
 import uuid
 import struct
-from edk2toollib.uefi.telem.cper_section_data import SECTION_PARSER_PLUGIN
+from edk2toollib.windows.telem.cper_section_data import SECTION_PARSER_PLUGIN
 
 """
 A MU Telemetry section has the following structure:
