+++ edk2toollib/windows/telem/friendlynames.py
